refactor: 이미지 카드 컴포넌트 분리
