+++ apps/app/src/screens/Tab/PastDiaryScreen.tsx
@@ -1,29 +1,15 @@
-import {FlatList, Image, StyleSheet, TouchableOpacity} from 'react-native';
+import {FlatList, StyleSheet} from 'react-native';
 import ScreenLayout from '../../components/@common/ScreenLayout';
-import Typography from '../../components/@common/Typography';
-import {screenWidth} from '../../constants/screenDimensions';
 import {formatDateToYYMMDD} from '../../utils/formatDate';
 import {PastDiaryImageResponseDto} from '../../types/dtos/responseDtos/PastDiary';
 import {mockPastDiaryImageList} from '../../constants/mockDatas/pastDiaryImageList';
-
-const CARD_MARGIN = 7;
-const CARD_WIDTH = (screenWidth - CARD_MARGIN * 3) / 2;
+import PastDiaryImageCard from '../../components/BottomTab/PastDiary/PastDiaryImageCard';
 
 function PastDiaryScreen() {
   const renderItem = ({item}: {item: PastDiaryImageResponseDto}) => {
     const formattedDate = formatDateToYYMMDD(item.createdAt);
 
-    return (
-      <TouchableOpacity style={pastDiaryScreenStyle.imageContainer}>
-        <Image
-          source={{uri: item.imageUrl}}
-          style={pastDiaryScreenStyle.image}
-        />
-        <Typography type={'headline-20'} style={pastDiaryScreenStyle.date}>
-          {formattedDate}
-        </Typography>
-      </TouchableOpacity>
-    );
+    return <PastDiaryImageCard imageUrl={item.imageUrl} date={formattedDate} />;
   };
 
   return (
@@ -52,20 +38,5 @@ const pastDiaryScreenStyle = StyleSheet.create({
   columnWrapper: {
     justifyContent: 'space-between',
     marginBottom: 7,
-  },
-  imageContainer: {
-    width: CARD_WIDTH,
-  },
-  image: {
-    width: '100%',
-    height: CARD_WIDTH,
-    borderRadius: 15,
-  },
-  date: {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: [{translateX: -CARD_WIDTH * 0.21}, {translateY: -15}],
-    color: 'rgba(255, 255, 255, 0.8)',
   },
 });
